+++ src/pptx/oxml/slide.py
@@ -17,6 +17,8 @@ from pptx.oxml.timing import (
     CT_SlideTiming,
     CT_TimeNodeList,
     CT_TLCommonTimeNodeData,
+    CT_TLTimeCondition,
+    CT_TLTimeConditionList,
     CT_TLTimeNodeParallel,
     CT_TLTimeNodeSequence,
 )
@@ -48,6 +50,8 @@ __all__ = [
     "CT_TimeNodeList",
     "CT_TLCommonTimeNodeData",
     "CT_TLMediaNodeVideo",
+    "CT_TLTimeCondition",
+    "CT_TLTimeConditionList",
     "CT_TLTimeNodeParallel",
     "CT_TLTimeNodeSequence",
     "CT_SideDirectionTransition",